Create README.md
By first examining standard fundus photos that have appeared in other DR  papers in the literature, I was able to collect a framework of images that exhibit early, middle and late BDR,  (Backround DR) and early and late PDR (Proliferative DR). I then proceeded to try to obtain histograms of these stage photos to compare them for  horizontal plot, (x axis) red, green, and blue  variation and vertical plot, (y axis) amplitude variation. Preliminary review of some pilot fundus photo/ histograms illustrated that the more advanced the DR the greater the horizontal and vertical plot alterations. Obviously, a much greater number of histograms and cumulative distribution plots with greater quantitative penetrance need to be completed, but this would appear to be an interesting hypothesis to explore as a signature of different stages of DR and the risk level. Also, one might hypothesize that not only the central and peripheral retina may need careful repeat fundus photos and histograms, but also where there might be false negatives, (the patient says my vision is down) but the fundus photo is fine or false positives (where the patient says my vision is fine, but the fundus photo is illustrating say an impending macula (central vision) defect. So fundus photos/histograms that would inform on false positives and false negatives also need to be generated. This would be the initial framework for this project.