Add files via upload
Vaishnavi et al was uploaded to introduce the reader to the complexities of fundus photo detail in studying the applications of AI in DR: Table 1: Analysis of DR based on the region method, Table 2:  Analysis of DR using pattern matching, Table 3: Analysis of DR with template matching and filtering/deconstruct--reconstruct e.g blood vessels  or optic nerve or macula or selected pixel segments of each anatomic location, Table 4: Watershed transformations to analyze optic disc, exudates, hemorrhages with 90%+ sensitivity. Each approach has advantages and disadvantages. So, for example, the watershed analysis to determine optic nerve lesions does that well, but differences between normal and pathologic were not included.
Thus, one might turn to a more simplified model for analysis such as in Kashyan et al IEEE Journal 2017 presented at 2nd Intl Conference of Convergence Technology. While the model illustrates a clear and less arduous method of examining the mean, median, and standard deviation of a normal vs and DR histogram and a careful measurement of "Euclidian distance between them" the relationships of structure and function, or lack thereof, not to mention stage or degree of difference between normal and pathology are far from clear.
It would appear that an all consuming over-riding principle for normal versus pathology in DR fundus photos might have to give way to a reasonable and likely statistical and practical method of simply calling a fundus photo in a risk level for subsequent progression.
Thus, with the risk level say above 80% from the fundus photo(s) single or serially, the innocent patient in a cloud on where his/her diabetes management stands could say to himself: "this is serious" or "not serious" at this time. This is one of the goals of this research. i.e. to serially monitor the diabetic fundus photo changes, develop a set of low, moderate and high risk parameters that will enable the diabetic patient to get some self assessment of his diabetic disease and chances for visual loss or compromise.